==> 4_FASTA-SMILES(m).ipynb <==
## Obtaining the FASTA or SMILES sequences for each CIDm chemical and corresponding proteins
After filtering the original dataset to only include CIDm chemicals and the corresponding binding proteins (see Notebook #3), it was necessary to map each of those ids to the respective SMILES or FASTA sequence. This information will be used later to calculate molecular feature descriptors for both types of molecules.

import pandas as pd

pc_binding = pd.read_table('pc_binding.tsv') 
pc_binding = pc_binding.drop(['Unnamed: 0'], axis=1) 

CIDm_binding = pc_binding.drop_duplicates('chemical') 
# takes binding pairs set and drops duplicate chemical IDs (goal is to have list of unique chemicals) 

CIDm_binding = CIDm_binding.drop(['protein', 'experimental'], axis=1) 
CIDm_binding = CIDm_binding.reset_index(drop=True) 
# drops all columns except chemical and resets index (IDs are already ordered) 

chemSMILES = pd.read_table('chemicals.v5.0.tsv.gz') 
# recall this is the master dataset of all chemicals and their info/SMILES for humans

len(chemSMILES) 

chemSMILES.head(5) 

chemSMILES = chemSMILES.drop(['name', 'molecular_weight'], axis=1) 
# keep only chemical and SMILES columns

chemSMILES.head(5) 

chemSMILES = chemSMILES[~chemSMILES['chemical'].str.contains("CIDs")] 
# filters out CIDs since our set is only CIDm (makes later merge faster) 

chemSMILES = chemSMILES.reset_index(drop=True) 
# reset index (already ordered) 

CIDm_binding = pd.merge(CIDm_binding, chemSMILES, on=['chemical']) 
# merge to get list of our unique chemicals and their SMILES

len(CIDm_binding) 
# 2 CIDm did not have a SMILES in the master list (?) 

CIDm_binding.columns = ['chemical', 'SMILES'] 
# rename columns

CIDm_binding.head() 

CIDm_binding.to_csv('CIDm_binding.tsv', sep='\t') 

protm_binding = pc_binding.drop_duplicates('protein') 
# gets list of unique proteins from binding pairs set

len(protm_binding) 

protm_binding = protm_binding.drop(['chemical', 'experimental'], axis=1) 
# keep only protein column

protm_binding = protm_binding.sort_values(by=['protein']) 
# sorts in accending order

protm_binding = protm_binding.reset_index(drop=True) 
# reset index

from Bio import SeqIO

with open('9606.protein.sequences.v10.5.fa') as fasta_file:
    protIDs = []
    sequences = []
    lengths = [] 
    for seq_record in SeqIO.parse(fasta_file, 'fasta'):
        protIDs.append(str(seq_record.id)) 
        sequences.append(str(seq_record.seq))
        lengths.append(len(seq_record.seq)) 
# parses FASTA file and sends output to lists (of IDs, sequences, and lengths of sequences) 

s1 = pd.Series(protIDs, name='protein')
s2 = pd.Series(sequences, name='FASTA')
s3 = pd.Series(lengths, name='length') 
# takes each list and turns it into a pandas series

protseq = pd.DataFrame(dict(protein=s1, FASTA=s2, length=s3)) 
# first turns the series into entries of a dictionary and then turns this dictionary into a dataframe with each of the series as a column

protm_binding = pd.merge(protm_binding, protseq, on=['protein']) 
# merge our unique proteins with master set with FASTA for all proteins in human body

protm_binding
# now we have our proteins and their FASTA

protm_binding = protm_binding.drop(['length'], axis=1) 
# remove length column

protm_binding

protm_binding.to_csv('protm_binding.tsv', sep='\t')
