==> 5_Compute_Chemical_Features.ipynb <==
## Computing Chemical Features using Mordred
This notebook uses the Mordred descriptor calculator to calculate 1613 physiological descriptors for each CIDm chemical

from rdkit import Chem
from mordred import Calculator, descriptors
import pandas as pd
import numpy as np
from matplotlib import pyplot as plt
%matplotlib inline

calc = Calculator(descriptors, ignore_3D=True) 
# creates descriptor calculator with all descriptors

print "Total number of 2D descriptors:", len(calc.descriptors) 

CIDm_binding = pd.read_csv('CIDm_binding.tsv', sep='\t')
CIDm_binding = CIDm_binding.drop(['Unnamed: 0'], axis=1) 

CIDm_binding.head()

for chunk in range(1,51):
    CIDm_tmp = CIDm_binding[(chunk-1)*5000:(chunk)*5000]
    smiles = list(CIDm_tmp['SMILES']) 
    chemicals = pd.DataFrame(list(CIDm_tmp['chemical']), columns=['chemical'])
    molecules = [Chem.MolFromSmiles(smi) for smi in smiles]
    cols = list(calc.pandas(molecules[100:110]).columns)
    
    dat = np.zeros([len(molecules), len(cols)])
    for mol in range(len(molecules)):
        if mol%100 == 0:
            print 'iteration number ' + str(mol)
        try:
            tmp_mds = np.asarray(calc(molecules[mol]),dtype=np.float64)
        except:
            tmp_mds = np.zeros([1,len(cols)])
            
        mds = np.zeros([1,len(tmp_mds)])
        for i in range(len(tmp_mds)-1):
            mds[0,i] = float(tmp_mds[i])
        dat[mol,:] = mds

    MDs = pd.DataFrame(columns=cols, data=dat)
    test = pd.concat([chemicals, MDs], axis=1)
    name = str(chunk) + '_CIDm_MDs.tsv'
    test.to_csv(name,sep='\t',index=False)


for chunk in range(52,53):
    CIDm_tmp = CIDm_binding[(chunk-1)*5000:(chunk)*5000]
    smiles = list(CIDm_tmp['SMILES']) 
    chemicals = pd.DataFrame(list(CIDm_tmp['chemical']), columns=['chemical'])
    molecules = [Chem.MolFromSmiles(smi) for smi in smiles]
    cols = list(calc.pandas(molecules[100:110]).columns)
    
    dat = np.zeros([len(molecules), len(cols)])
    for mol in range(len(molecules)):
        if mol%100 == 0:
            print 'iteration number ' + str(mol)
        try:
            tmp_mds = np.asarray(calc(molecules[mol]),dtype=np.float64)
        except:
            tmp_mds = np.zeros([1,len(cols)])
            
        mds = np.zeros([1,len(tmp_mds)])
        for i in range(len(tmp_mds)-1):
            mds[0,i] = float(tmp_mds[i])
        dat[mol,:] = mds

    MDs = pd.DataFrame(columns=cols, data=dat)
    test = pd.concat([chemicals, MDs], axis=1)
    name = str(chunk) + '_CIDm_MDs.tsv'
    test.to_csv(name,sep='\t',index=False)
